#64 Refactored casefiles, docs, notes into separate components
