+++ angular-client/src/app/components/participants/participant-profile/notes/notes.component.css
@@ -1,0 +1,1 @@
+@import '../participant-profile.component.css'
